% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malayrootwords.R
\docType{data}
\name{malayrootwords}
\alias{malayrootwords}
\title{Data of Malay root words}
\format{
A tibble with 4320 rows and 2 variables:
\describe{
\item{\verb{Col Word}}{dbl Malay Word}
\item{\verb{Root Word}}{dbl Malay Root Word}
}
}
\usage{
malayrootwords
}
\description{
Data of Malay root words
}
\keyword{datasets}
