% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaysia_politic_sentiment.R
\docType{data}
\name{malaysia_politic_sentiment}
\alias{malaysia_politic_sentiment}
\title{Malaysia Politic Tweets Sentiment Dataset (Positive, Negative or Neutral)}
\format{
A tibble with 71 rows and 3 variables:
\describe{
\item{\code{id}}{dbl Represents a unique identifier assigned to each tweet}
\item{\code{text}}{dbl Tweet related to Malaysia politics}
\item{\code{Sentiment}}{dbl The sentiment classification assigned to each tweet}
}
}
\usage{
malaysia_politic_sentiment
}
\description{
Malaysia Politic Tweets Sentiment Dataset (Positive, Negative or Neutral)
}
\keyword{datasets}
