% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malaystopwords}
\alias{malaystopwords}
\title{Malay Stop Words.}
\format{
A data frame with one variable: \code{stopwords}
}
\usage{
malaystopwords
}
\description{
Malay Stop Words List from Abdullah, M. T. (2005). Improvement of Malay information retrieval using local stop words. Proc. IATC, 2005, 1-7.
}
\keyword{datasets}
