% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.R
\docType{data}
\name{normalized}
\alias{normalized}
\title{Data of Malay root words}
\format{
A tibble with 153 rows and 2 variables:
\describe{
\item{\verb{Col Word}}{dbl Word}
\item{\verb{Normalized Word}}{dbl Normalized Word}
}
}
\usage{
normalized
}
\description{
Data of Malay root words
}
\keyword{datasets}
