% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_url}
\alias{remove_url}
\title{Remove URL links}
\usage{
remove_url(string)
}
\arguments{
\item{string}{String to change}
}
\value{
Returns a string with URL links removed
}
\description{
Remove URL links
}
\details{
\code{remove_url()} is an approach to remove link(s) from a string
}
\examples{
x <- c("test https://t.co/fkQC2dXwnc", "another one https://www.google.com/ to try")
remove_url(x)
}
