% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_general.R
\docType{data}
\name{sentiment_general}
\alias{sentiment_general}
\title{Data of Sentiment Words (Positive or Negative)}
\format{
A tibble with 1424 rows and 2 variables:
\describe{
\item{\code{Word}}{dbl Sentiment Word}
\item{\verb{Root Word}}{dbl Sentiment}
}
}
\usage{
sentiment_general
}
\description{
Data of Sentiment Words (Positive or Negative)
}
\keyword{datasets}
