% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaystopwords.R
\docType{data}
\name{malaystopwords}
\alias{malaystopwords}
\title{Data of Malay stop words}
\format{
A tibble with 512 rows and 1 variable:
\describe{
\item{\code{stopwords}}{dbl Malay stop words}
}
}
\usage{
malaystopwords
}
\description{
Data of Malay stop words
}
\keyword{datasets}
